Add context to request
The way go handles request specific variables is via a tree of contexts.

That means stuff like logging the IP for any log message is impossible because there's no way to get the context in the log formatter, which is the only place I can see to add custom fields *after* the log item is created.